+++ domains/money/src/main/kotlin/de/jfachwert/money/Geldbetrag.kt
@@ -798,7 +798,7 @@ open class Geldbetrag @JvmOverloads constructor(betrag: Number, currency: Curren
      *
      * @return z.B. "$19"
      */
-    fun toShortString(): String {
+    override fun toShortString(): String {
         return format("$#")
     }
 
@@ -818,7 +818,7 @@ open class Geldbetrag @JvmOverloads constructor(betrag: Number, currency: Curren
      *
      * @return z.B. "19.0012 USD"
      */
-    fun toLongString(): String {
+    override fun toLongString(): String {
         val formatter = DecimalFormat.getInstance()
         formatter.minimumFractionDigits = context.maxScale
         formatter.minimumFractionDigits = context.maxScale